don't kill WANDERER ('good' or 'neutral' opponents), defeat them

+++ src/server/mon-util.c
@@ -1139,6 +1139,13 @@ static void player_kill_monster(struct player *p, struct chunk *c, struct source
                 msgt(p, soundfx, "You have killed %s.", m_name);
             }
 
+            // WANDERER ('good' or 'neutral' monster)
+            else if (rf_has(mon->race->flags, RF_WANDERER))
+            {
+                msg_format_near(p, MSG_GENERIC, " has defeated %s.", m_name);
+                msgt(p, soundfx, "You have defeated %s.", m_name);
+            }
+
             /* Death by physical attack -- unusual monster */
             else if (monster_is_destroyed(mon->race))
             {
@@ -1168,10 +1175,18 @@ static void player_kill_monster(struct player *p, struct chunk *c, struct source
             switch (note)
             {
                 case MON_MSG_DESTROYED:
-                    msg_format_complex_near(p, MSG_GENERIC, "%s is destroyed.", name);
+                    // WANDERER ('good' or 'neutral' monster)
+                    if (rf_has(mon->race->flags, RF_WANDERER))
+                        msg_format_complex_near(p, MSG_GENERIC, "%s is defeated.", name);
+                    else
+                        msg_format_complex_near(p, MSG_GENERIC, "%s is destroyed.", name);
                     break;
                 default:
-                    msg_format_complex_near(p, MSG_GENERIC, "%s dies.", name);
+                    // WANDERER ('good' or 'neutral' monster)
+                    if (rf_has(mon->race->flags, RF_WANDERER))
+                        msg_format_complex_near(p, MSG_GENERIC, "%s defeated.", name);
+                    else
+                        msg_format_complex_near(p, MSG_GENERIC, "%s dies.", name);
             }
             add_monster_message(p, mon, note, true);
         }
@@ -1186,14 +1201,22 @@ static void player_kill_monster(struct player *p, struct chunk *c, struct source
             type = MSG_BROADCAST_KILL_KING;
 
         /* Give credit to the killer */
-        strnfmt(buf, sizeof(buf), "%s was slain by %s %s.", mon->race->name, title, p->name);
+        // WANDERER ('good' or 'neutral' monster)
+        if (rf_has(mon->race->flags, RF_WANDERER))        
+            strnfmt(buf, sizeof(buf), "%s was defeated by %s %s.", mon->race->name, title, p->name);
+        else
+            strnfmt(buf, sizeof(buf), "%s was slain by %s %s.", mon->race->name, title, p->name);
 
         /* Tell every player (including the killer because it's easy to miss in message window) */
         msg_broadcast(p, buf, type);
         msg_print(p, buf, type);
 
         /* Message for event history */
-        strnfmt(logbuf, sizeof(logbuf), "Killed %s", mon->race->name);
+        // WANDERER ('good' or 'neutral' monster)
+        if (rf_has(mon->race->flags, RF_WANDERER))        
+            strnfmt(logbuf, sizeof(logbuf), "Defeated %s", mon->race->name);
+        else
+            strnfmt(logbuf, sizeof(logbuf), "Killed %s", mon->race->name);
 
         /* Record this kill in the event history */
         history_add_unique(p, logbuf, HIST_SLAY_UNIQUE);
@@ -1209,7 +1232,11 @@ static void player_kill_monster(struct player *p, struct chunk *c, struct source
             if (party_share_with(p, p->party, q) && mflag_has(q->mflag[mon->midx], MFLAG_HURT))
             {
                 /* Message for event history */
-                strnfmt(logbuf, sizeof(logbuf), "Helped to kill %s", mon->race->name);
+                // WANDERER ('good' or 'neutral' monster)
+                if (rf_has(mon->race->flags, RF_WANDERER))     
+                    strnfmt(logbuf, sizeof(logbuf), "Helped to defeat %s", mon->race->name);
+                else
+                    strnfmt(logbuf, sizeof(logbuf), "Helped to kill %s", mon->race->name);
 
                 /* Record this kill in the event history */
                 history_add(q, logbuf, HIST_HELP_UNIQUE);
@@ -1240,7 +1267,11 @@ static void player_kill_monster(struct player *p, struct chunk *c, struct source
     {
         int drain = 1 + (mon->level / 2) + p->lev * 4 / 5;
         if (drain > mon->maxhp) drain = mon->maxhp;
-        msg(p, "You absorb the life of the dying soul.");
+        // WANDERER ('good' or 'neutral' monster)
+        if (rf_has(mon->race->flags, RF_WANDERER))  
+            msg(p, "You absorb the morale of the fleeing opponent.");
+        else
+            msg(p, "You absorb the life of the dying soul.");
         hp_player_safe(p, 1 + drain / 2);
     }
